+++ src/util/formatDate.js
@@ -1,4 +1,4 @@
-import { format } from 'date-fns';
+import { format, parseISO, formatISO } from 'date-fns';
 import { nb as locale } from 'date-fns/locale';
 
 const o = { locale };
@@ -12,3 +12,6 @@ export function periodDates({ startDate, endDate }, strategy = dateLong) {
   if (!startDate || !endDate) throw new Error('Invalid period object');
   return [startDate.toDate(), endDate.toDate()].map(strategy).join(' – ');
 }
+
+export const parseDate = d => parseISO(d);
+export const formatISOShort = d => formatISO(d, { representation: 'date' });