+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Added
 
+- Filter KPI-progression
 - E2E-tests (alpha)
 
 ### Changes